requirements.txt
descartes==1.1.0
matplotlib==3.5.1
numpy==1.21.5
opencv_python==4.6.0.66
Pillow==9.4.0
scikit_image==0.19.3
Shapely==1.8.4
skimage==0.0
